Add note :NO-FLOATING-POINT-UNDERFLOW-BY-DEFAULT
This note assumes that the implementation purposefully does not conform to
12.1.4.3 by not signaling FLOATING-POINT-UNDERFLOW by default. (The
implementation should have an option of toggling this behaviour off in order
to be able to signal this condition when underflows occur.)

A reason for lack of conformance might be contradictory wording in the spec:
12.1.4.3 mentions that the errors should be signaled, while arithmetic
operations mention that they "might signal" the error.

Tests EXP{T,}.{8,9,10,11} have been annotated with this note.