Add SUBTYPEP.CONS.44
We add a SUBTYPEP test that checks SUBTYPEP relationship for CONS types
with SATISFIES types in their CARs and CDRs.

Whether t1 is a subtype of t2 depends on a series of RANDOM calls at
typechecking time - sometimes it might be a subtype, it might not.

This test is added because I see SBCL and CCL failures in this case.

In addition to the test, we define four functions named
subtypep.cons.44.{foo,bar,baz,qux}. This is because the CLHS states that the
type specifier SATISFIES "denotes the set of all objects that satisfy the
predicate PREDICATE-NAME, which must be a symbol whose global function
definition is a one-argument predicate."