Set the *random-state* for cmucl
Set the *random-state* to a known state when running these tests so
that the tests that use the random number generator always gets the
same state (provided the tests are run in the same order and number of
tests).

Add a few comments too on why we set the customizations to the values
they're given.