Check for symbol length in character.1.body
Cmucl converts a string to Unicode NFC form before creating a symbol.
This means some symbols have names that longer than 1 character.  The
test assumes the name has one character which causes an error in
cmucl.

This just adds a check that the length is 1 before extracting the
character from the symbol name.  This doesn't change behavior of the
test (which was already assuming a length of 1), but allows cmucl not
to signal an error but to finish the test with a meaningful list of
characters whose corresponding symbol name is more than one character
long.

A couple of minor tweaks:
- Ignore sse2f fasl created by cmucl/x86.
- Add comment to makefile on how to run with cmucl.