Fix #49:  Use rational instead of rationalize in format.e.26
In the test `format.e.26`, the float `x` is converted to a rational
number via `rationalize` so that the correctly rounded result can be
computed.  However `rationalize` can produce a rational that is very
close ot the float `x` but is slightly different.  This can produce
incorrect rounding since the value being rounded is different from the
actual float.

Thus use `rational` to find the exact rational value of `x`.